debug bad

+++ brown_and_dennis.py
@@ -7,21 +7,23 @@ class BrownAndDennis:
         self.n = 4
         self.m = m
         assert self.m >= self.n
-        self.x_0 = np.array([25, 5, -5, -1])
+        self.x_0 = np.array([25, 5, -5, -1], dtype="float32")
         self.f_minimun = 85822.2
 
     def func(self, x):
         f = 0
         for l in range(self.m):
-            t = l / 5
+            t = l / 5.
             f = f + (x[0] + t * x[1] -
                      np.exp(t))**2 + (x[2] + x[3] * np.sin(t) - np.cos(t))**2
         return f
 
     def grad(self, x):
         g = np.zeros_like(x, dtype="float32")
+        # import pdb
+        # pdb.set_trace()
         for l in range(self.m):
-            t = l / 5
+            t = l / 5.
             g[0] += 2 * (x[0] + t * x[1] - np.exp(t))
             g[1] += 2 * (x[0] + t * x[1] - np.exp(t)) * t
             g[2] += 2 * (x[2] + x[3] * np.sin(t) - np.cos(t))
